Organized UnitTests into respective folders
