+++ dotnet/WebTestFramework/Framework/UnitTests/Elements/TextFieldTests.cs
@@ -1,7 +1,8 @@
 ﻿using Framework.Elements;
 using Framework.UnitTests.PageObjects;
 using NUnit.Framework;
-namespace Framework.UnitTests
+
+namespace Framework.UnitTests.Elements
 {
     [TestFixture]
     [Parallelizable]
