+++ dotnet/WebTestFramework/Framework/UnitTests/PageObjects/NavigationMenuTests.cs
@@ -1,10 +1,9 @@
 ﻿using Framework.Elements;
 using Framework.Interfaces;
-using Framework.UnitTests.PageObjects;
 using NUnit.Framework;
 using OpenQA.Selenium;
 
-namespace Framework.UnitTests
+namespace Framework.UnitTests.PageObjects
 {
     [TestFixture]
     [Parallelizable]
